refactor: extract zenyuv crate from zenjpeg rgb_to_yuv module
New workspace member `zenyuv/` with:
- `#![no_std]`, `#![forbid(unsafe_code)]`, MIT/Apache-2.0
- All AVX2 and magetypes-generic kernels moved from zenjpeg
- Parity tests against yuv crate (dev-dep only)
- Doctest for the public API

zenjpeg's `color::rgb_to_yuv` is now a thin re-export of zenyuv.
Bench uses zenyuv directly. Performance unchanged (14.7µs vs 16.5µs
at 256px, same numbers as before extraction).

zenyuv is designed to eventually replace the `yuv` crate dependency
across zenjpeg, zenwebp, and zenavif once it covers:
- YCbCr→RGB (decode direction)
- BT.709/2020 matrices, limited range
- 10/12/16-bit depth, 4:2:2 subsampling
- Sharp YUV (iterative perceptual chroma)
